feat: fix the ordering of tables in plan_test and move cycle detection back to original package
